\docType{package}
\name{seedGerm-package}
\alias{seedGerm-package}
\title{Seed germination outines}
\description{
  This package is a collection of different seed
  germination routines.
}
\details{
  \tabular{ll}{ Package: \tab seeGerm\cr Type: \tab
  Package\cr Version: \tab 0.1.0\cr Date: \tab
  2013-10-01_10-23\cr License: \tab GPL (>= 2)\cr LazyLoad:
  \tab yes\cr }
}
\author{
  Rainer M Krug \email{Rainer@krugs.de}
}

